refactor: rename options to payload
