+++ packages/app-bridge/src/dispatch/NavigationManager.ts
