+++ packages/app-bridge/src/index.ts
@@ -8,7 +8,7 @@ export * from './AppBridgeBase';
 export * from './AppBridgeBlock';
 export * from './AppBridgeCreateAsset';
 export * from './AppBridgeTheme';
-export * from './commands';
+export * from './dispatch';
 export * from './errors';
 export * from './react';
 export * from './tests';
